fix: rename all hook
