+++ src/pages/collection/CollectionEdit.tsx
@@ -1,5 +1,5 @@
-import CollectionMange from '../../components/CollectionMange';
-import useColedit from '../../hooks/useColedit';
+import CollectionMange from "../../components/CollectionMange";
+import useCollectionEdit from "../../hooks/useCollectionedit";
 export const Collection = () => {
   const {
     formData,
@@ -11,24 +11,21 @@ export const Collection = () => {
     onChangeDate,
     onChangeRemindDate,
     handleAddNotice,
-    handleDeleteNotice
-  } = useColedit();
+    handleDeleteNotice,
+  } = useCollectionEdit();
   return (
-
-      <CollectionMange
-        formData={formData}
-        notices={notices}
-        handleChange={handleChange}
-        handleNoticeChange={handleNoticeChange}
-        handleSave={handleSave}
-        handleReset={handleReset}
-        onChangeDate={onChangeDate}
-        onChangeRemindDate={onChangeRemindDate}
-        handleAddNotice={handleAddNotice}
-        handleDeleteNotice={handleDeleteNotice}
-      />
-
+    <CollectionMange
+      formData={formData}
+      notices={notices}
+      handleChange={handleChange}
+      handleNoticeChange={handleNoticeChange}
+      handleSave={handleSave}
+      handleReset={handleReset}
+      onChangeDate={onChangeDate}
+      onChangeRemindDate={onChangeRemindDate}
+      handleAddNotice={handleAddNotice}
+      handleDeleteNotice={handleDeleteNotice}
+    />
   );
 };
 export default Collection;
-
